Machine learning model 1.0
This version crashes when I try to run the second cell to execute. Be aware of the data I take, that's a self made csv file called sgen_table, this one is also included. This was solely for the purpose of having a not to big dataset compared to the load dataset